Begin decoupling of moving parts.
The entire datepicker widget is a construct of several moving parts. The sum of
which create the commonly accepted form of a date picker.

Reflex enables tremendous modularity due to the reliable composition of
individual parts. Thus I am leveraging that so that users of this package would
be able to pick and choose their required moving parts to suit their needs. Or
just use a baked in default with a few tweaks.