Decoupling ad infinitum & styling
Most components of this widget have been decoupled and placed in their own
modules. There is a separate 'Styles' module that contains the defaults and some
basic wrappers for the various bits, designed to work with the generic CSS
provided in the given file.

The widget now doesn't look completely awful and should have some flexibility
with respect to styling. Unsure of how to ease the pain of including the layout
CSS but that's a different problem...

The decoupling might have been taken a bit far as there is a simple gap that can
cause the updates to not work as intended. Unless you combine the date set value
and day selected events when constructing the core, you may miss out on updates.